+++ CHANGES.rst
@@ -10,6 +10,8 @@ Unreleased
 -   Remove usage of ``warnings.catch_warnings``. :issue:`2690`
 -   Remove ``max_form_parts`` restriction from standard form data parsing and only use
     if for multipart content. :pr:`2694`
+-   ``Response`` will avoid converting the ``Location`` header in some cases to preserve
+    invalid URL schemes like ``itms-services``. :issue:`2691`
 
 
 Version 2.3.3